Data Warehousing and Business Intelligence
The net sale, revenues of product category, GDP of regions and gross prot of Adidas is analyzed in this project for different regions and different time frames like quarterly, yearly. The main feature of this report is the analyzing the net sale of Adidas for ten years (2008 to 2017) of time span. Gathering the data and extracting it, transformation of the data, loading it in the database and deploying it into a cube using full automation is shown in this report. The main objective is to analyze the net sale of Adidas in order to form a better marketing and business strategy in the region where it is lacking behind to compete with it others competitors. By studying these data it would help Adidas on enhancing its sale worldwide. The net sale by product categories can also be analyzed and further it can be increased.